moved ImageManipulator
* added `expo-errors`
